Se prueba y finaliza product-service

+++ sales_management/product-service/src/main/java/com/salesmanagementplatform/product/repository/ProductStockRepository.java
@@ -3,6 +3,6 @@ package com.salesmanagementplatform.product.repository;
 import com.salesmanagementplatform.product.model.ProductStockModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductStockRepository extends JpaRepository<ProductStockModel, Integer> {
+public interface ProductStockRepository extends JpaRepository<ProductStockModel, Long> {
     ProductStockModel findByProduct_productNumber(int productNumber);
 }
